# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IncidentHub API
  version: 1.0.0
  description: |
    Production-style incident management API built with Node.js, Express, PostgreSQL, and Prisma.
servers:
  - url: http://localhost:5001/api
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: IncidentHub Backend
        description:
          type: string
          example: Main backend API for incident tracking
    CreateTicketRequest:
      type: object
      required: [title, description, projectId]
      properties:
        title:
          type: string
          example: Login API returns 500
        description:
          type: string
          example: Users receive internal server error while logging in.
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
          example: OPEN
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
          example: HIGH
        projectId:
          type: integer
          example: 1
        assigneeId:
          type: integer
          example: 1
...
paths:
...
  /projects:
    post:
      tags: [Projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      tags: [Projects]
      summary: Get all projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects fetched successfully
        '401':
          description: Unauthorized
...
  /tickets:
    post:
      tags: [Tickets]
      summary: Create a ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket created successfully

    get:
      tags: [Tickets]
      summary: Get all tickets with filters, search, sorting, and pagination
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        - in: query
          name: priority
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: assigneeId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, updatedAt, priority, status, title]
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Tickets fetched successfully
...